+++ requirements.txt
@@ -1,185 +1,9 @@
-absl-py==0.15.0
-aplus==0.11.0
-appdirs==1.4.4
-appnope==0.1.0
-astropy==4.0.1.post1
-astunparse==1.6.3
-attrs==19.3.0
-backcall==0.1.0
-bleach==3.1.4
-boto3==1.14.12
-botocore==1.17.12
-bqplot==0.12.13
-branca==0.3.1
-cached-property==1.5.2
-cachetools==4.1.1
-certifi==2020.4.5.1
-chardet==3.0.4
-clang==5.0
-click==7.1.2
-cloudpickle==1.4.1
-cmake==3.22.0
-cmdstanpy==0.4.0
-convertdate==2.2.1
-cycler==0.10.0
-Cython==0.29.20
-dask==2.19.0
-dataclasses==0.8
-decorator==4.4.2
-defusedxml==0.6.0
-distlib==0.3.1
-docutils==0.15.2
-entrypoints==0.3
-ephem==3.7.7.1
-fbprophet==0.6
-filelock==3.0.12
-flatbuffers==1.12
-future==0.18.2
-gast==0.4.0
-google-auth==1.35.0
-google-auth-oauthlib==0.4.6
-google-pasta==0.2.0
-grpcio==1.42.0
-h5py==3.1.0
-holidays==0.10.2
-idna==2.9
-importlib-metadata==4.8.2
-importlib-resources==5.1.0
-ipydatawidgets==4.0.1
-ipykernel==5.2.0
-ipyleaflet==0.13.0
-ipympl==0.5.6
-ipython==7.13.0
-ipython-genutils==0.2.0
-ipyvolume==0.5.2
-ipyvue==1.3.3
-ipyvuetify==1.4.0
-ipywebrtc==0.5.0
-ipywidgets==7.5.1
-jedi==0.16.0
-Jinja2==2.11.1
-jmespath==0.10.0
-joblib==0.14.1
-jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client==6.1.2
-jupyter-console==6.1.0
-jupyter-contrib-core==0.3.3
-jupyter-contrib-nbextensions==0.5.1
-jupyter-core==4.6.3
-jupyter-highlight-selected-word==0.2.0
-jupyter-latex-envs==1.4.6
-jupyter-nbextensions-configurator==0.4.1
-keras==2.6.0
-Keras-Preprocessing==1.1.2
-keyring==21.2.0
-kiwisolver==1.2.0
-korean-lunar-calendar==0.2.1
-llvmlite==0.33.0
-LunarCalendar==0.0.9
-lxml==4.6.4
-Markdown==3.3.6
-MarkupSafe==1.1.1
-matplot==0.1.9
-matplotlib==3.2.1
-mistune==0.8.4
-multitasking==0.0.10
-nbconvert==5.6.1
-nbformat==5.0.5
-nest-asyncio==1.3.3
-nltk==3.5
-notebook==6.0.3
-numba==0.50.1
-numpy==1.19.5
-oauthlib==3.1.0
-opt-einsum==3.3.0
-pandas==1.0.3
-pandocfilters==1.4.2
-parso==0.6.2
-patsy==0.5.1
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==7.1.2
-pkginfo==1.5.0.1
-progressbar2==3.51.4
-prometheus-client==0.7.1
-prompt-toolkit==3.0.5
-protobuf==3.19.1
-psutil==5.7.0
-ptyprocess==0.6.0
-pyarrow==0.17.1
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-Pygments==2.6.1
-pyloco==0.0.139
-PyMeeus==0.3.7
-pyparsing==2.4.7
-pyrsistent==0.16.0
-PySocks==1.7.1
-pystan==2.19.1.1
-python-dateutil==2.8.1
-python-utils==2.4.0
-pythreejs==2.2.0
+
 pytrends==4.7.3
-pytz==2019.3
-PyYAML==5.3.1
-pyzmq==19.0.0
-qtconsole==4.7.2
-QtPy==1.9.0
-readme-renderer==25.0
-regex==2020.11.13
-requests==2.23.0
-requests-oauthlib==1.3.0
-requests-toolbelt==0.9.1
-rsa==4.7.2
-s3fs==0.2.2
-s3transfer==0.3.3
-scikit-learn==0.22.2.post1
-scipy==1.4.1
-seaborn==0.10.1
-Send2Trash==1.5.0
-setuptools-git==1.2
-SimpleWebSocketServer==0.1.1
-six==1.15.0
-statsmodels==0.11.1
-tabulate==0.8.7
-tensorboard==2.6.0
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.0
-tensorflow==2.6.2
-tensorflow-estimator==2.6.0
-termcolor==1.1.0
-terminado==0.8.3
-testpath==0.4.4
-textblob==0.15.3
-toolz==0.10.0
-tornado==6.0.4
-tqdm==4.45.0
-traitlets==4.3.3
-traittypes==0.2.1
-tweepy==3.9.0
-twine==3.1.1
-typing==3.7.4.1
-typing-extensions==3.7.4.3
-urllib3==1.25.8
-ushlex==0.99.1
-vaex==3.0.0
-vaex-arrow==0.5.1
-vaex-astro==0.7.0
-vaex-core==2.0.3
-vaex-hdf5==0.6.0
-vaex-jupyter==0.5.2
-vaex-ml==0.9.0
-vaex-server==0.3.1
-vaex-viz==0.4.0
-virtualenv==20.4.2
-wcwidth==0.1.9
-webencodings==0.5.1
-websocket-client==0.57.0
-Werkzeug==2.0.2
-widgetsnbextension==3.5.1
-wrapt==1.12.1
-xarray==0.15.1
-xgboost==1.5.1
-yfinance==0.1.67
-zipp==3.1.0
+pystan==2.19.1.1
+scikit-learn
+pandas
+numpy
+prophet==1.0.1
+yfinance==0.1.70
+
